Fix broken format strings

+++ src/subgraph/concatenate.c
@@ -326,8 +326,8 @@ enum xnn_status xnn_define_concatenate_impl(
 
   if (num_inputs > XNN_MAX_OPERATOR_OBJECTS) {
     xnn_log_error(
-      "failed to define %s operator with %zu inputs: number of inputs (%zu) exceeds the supported maximum (%zu)",
-      xnn_node_type_to_string(node_type), num_inputs, num_inputs, (size_t) XNN_MAX_OPERATOR_OBJECTS);
+      "failed to define %s operator with %d inputs: number of inputs (%d) exceeds the supported maximum (%d)",
+      xnn_node_type_to_string(node_type), num_inputs, num_inputs, XNN_MAX_OPERATOR_OBJECTS);
     return xnn_status_invalid_parameter;
   }
